wip: switch from regression to classification
This commit contains two strategies for picking view slots that ended up
not working.  These can be found in `_classification.py.bkup` and
`classification.py`.  I'm about to delete all of this code, but I wanted
to keep it available in the project history first.

This commit also refactors the transformation-prediction dataset module,
and includes updated tests.