fix: initialize InfiniteSampler with a valid epoch
I originally avoided doing this, so I'd get an error if the sampler was
used before `set_epoch()` was called.  Unfortunately, it turns out that
Lightning does draw examples from the data loader before setting the
epoch.  This occurs when Lightning is "checking that the data loader is
iterable".  I don't think the resulting training examples are actually
used for anything, though.

For reasons I don't understand, this is only a problem when using
multiple data loader processes.  With only a single data loader,
Lightning still appears to do this check, but apparently doesn't require
the sampler to be invoked somehow.