fix: spawn (rather than fork) dataloader worker processes
This seems to fix the race conditions/deadlocks/abort signals that have
been causing problems for the last few days.  I still don't know exactly
what the issue is, but spawning brand new processes is generally much
safer than forking the current process, and the memory/startup time
penalties aren't significant in the context of such a long-running
program.

In order to make this change, I had to move the code for copying the
origins database to /tmp from inside the worker processes to outside.