feat: use persistent worker subprocesses
Lightning recommends enabling the perseitent worker setting when using
"spawn"-style subprocesses, because these subprocesses take longer to
start and this setting reduces the number of times that needs to happen.

I enabled this setting partially to follow Lightning's recommendation
(i.e. to silence a warning message) and also partially because I was
experiencing deadlocks.  This temporarily seemed to avoid the deadlock,
but I really doubt it addressed the underlying problem, so I expect I'll
still have to deal with this eventually.