fix: don't allow non-finite sigma values
Prior to this commit, I was using a bespoke method of generating samples
from a Gaussian distribution.  On two separate occasions, this method
was responsible for weird bugs stemming from its ability to generate
non-finite values.  While I think I've now encountered every non-finite
value my method could generate, I decided that it was a mistake to use a
bespoke method in the first place, so I switched to the NumPy
implementation.  This requires some machinery to conveniently batch
NumPy `Generator` objects, which is why I didn't take this approach in
the first place, but it's not too bad.